lab4/user-service/requirements.txt
fastapi
uvicorn
asyncpg
python-jose[cryptography]
passlib
python-multipart
pyjwt